restart host operator using rollout
